# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Taiwan Tourism Open Data API
  description: '

    以 **交通部觀光署「景點 - 觀光資訊資料庫」** 政府開放資料（6,000+ 筆全台景點）為基礎的 RESTful API。


    資料來源：<https://data.gov.tw/dataset/7777>（觀光資料標準 V2.1）


    ### 資源 (Resources)

    | Resource | 說明 |

    |---|---|

    | `attractions` | 景點（支援關鍵字搜尋、多條件篩選、排序、分頁、附近景點查詢） |

    | `attractions/{id}/images` | 景點圖片（子資源） |

    | `categories` | 景點類型（文化類、生態類…） |

    | `cities` / `towns` | 縣市 / 鄉鎮市區 |

    | `reviews` | 遊客評論（1~5 星） |

    | `trips` / `trips/{id}/items` | 旅遊行程與行程項目 |

    | `stats` | 統計資訊（唯讀） |


    ### 驗證

    所有 **讀取 (GET)** 操作公開；**新增 / 修改 / 刪除 (POST / PUT / PATCH / DELETE)** 需在 Header
    帶入

    `X-API-Key: ntub-iot-2026`（可按右上角 **Authorize** 按鈕設定）。


    ### 慣例

    * 列表回應格式：`{"total", "page", "page_size", "pages", "items"}`，並附 `X-Total-Count`
    header

    * 建立成功回傳 `201 Created` 與 `Location` header；刪除成功回傳 `204 No Content`

    * 錯誤格式：`{"detail": "錯誤說明"}`（404 查無資料、409 資料衝突、422 驗證失敗）

    '
  contact:
    name: NTUB IoT App 作業
  license:
    name: 政府資料開放授權條款－第1版
    url: https://data.gov.tw/license
  version: 1.0.0
paths:
...
  /api/v1/cities:
    get:
      tags:
      - Cities 縣市
      summary: 列出縣市
      operationId: list_cities_api_v1_cities_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 名稱關鍵字，例如：臺北
          title: Q
        description: 名稱關鍵字，例如：臺北
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 頁碼（從 1 開始）
          default: 1
          title: Page
        description: 頁碼（從 1 開始）
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 每頁筆數（1~100）
          default: 20
          title: Page Size
        description: 每頁筆數（1~100）
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_CityOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Cities 縣市
      summary: 新增縣市
      operationId: create_city_api_v1_cities_post
      security:
      - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityOut'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cities/{city_code}:
    get:
      tags:
      - Cities 縣市
      summary: 取得單一縣市
      operationId: get_city_api_v1_cities__city_code__get
      parameters:
      - name: city_code
        in: path
        required: true
        schema:
          type: string
          title: City Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Cities 縣市
      summary: 整筆更新縣市
      operationId: replace_city_api_v1_cities__city_code__put
      security:
      - APIKeyHeader: []
      parameters:
      - name: city_code
        in: path
        required: true
        schema:
          type: string
          title: City Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Cities 縣市
      summary: 部分更新縣市
      operationId: update_city_api_v1_cities__city_code__patch
      security:
      - APIKeyHeader: []
      parameters:
      - name: city_code
        in: path
        required: true
        schema:
          type: string
          title: City Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Cities 縣市
      summary: 刪除縣市（連同其鄉鎮市區）
      operationId: delete_city_api_v1_cities__city_code__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: city_code
        in: path
        required: true
        schema:
          type: string
          title: City Code
      - name: force
        in: query
        required: false
        schema:
          type: boolean
          description: 若仍有景點使用此行政區，是否強制刪除（景點的行政區欄位會被清空）
          default: false
          title: Force
        description: 若仍有景點使用此行政區，是否強制刪除（景點的行政區欄位會被清空）
      responses:
        '204':
          description: Successful Response
        '409':
          description: 仍有景點位於此縣市且未指定 force=true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cities/{city_code}/towns:
    get:
      tags:
      - Cities 縣市
      summary: 列出縣市底下的鄉鎮市區
      operationId: list_city_towns_api_v1_cities__city_code__towns_get
      parameters:
      - name: city_code
        in: path
        required: true
        schema:
          type: string
          title: City Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TownOut'
                title: Response List City Towns Api V1 Cities  City Code  Towns Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/towns:
    get:
      tags:
      - Towns 鄉鎮市區
      summary: 列出鄉鎮市區
      operationId: list_towns_api_v1_towns_get
      parameters:
      - name: city_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 依縣市代碼篩選
          title: City Code
        description: 依縣市代碼篩選
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 名稱關鍵字
          title: Q
        description: 名稱關鍵字
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 頁碼（從 1 開始）
          default: 1
          title: Page
        description: 頁碼（從 1 開始）
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 每頁筆數（1~100）
          default: 20
          title: Page Size
        description: 每頁筆數（1~100）
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_TownOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Towns 鄉鎮市區
      summary: 新增鄉鎮市區
      operationId: create_town_api_v1_towns_post
      security:
      - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TownCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TownOut'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/towns/{town_code}:
    get:
      tags:
      - Towns 鄉鎮市區
      summary: 取得單一鄉鎮市區
      operationId: get_town_api_v1_towns__town_code__get
      parameters:
      - name: town_code
        in: path
        required: true
        schema:
          type: string
          title: Town Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TownOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Towns 鄉鎮市區
      summary: 整筆更新鄉鎮市區
      operationId: replace_town_api_v1_towns__town_code__put
      security:
      - APIKeyHeader: []
      parameters:
      - name: town_code
        in: path
        required: true
        schema:
          type: string
          title: Town Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TownCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TownOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Towns 鄉鎮市區
      summary: 部分更新鄉鎮市區
      operationId: update_town_api_v1_towns__town_code__patch
      security:
      - APIKeyHeader: []
      parameters:
      - name: town_code
        in: path
        required: true
        schema:
          type: string
          title: Town Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TownUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TownOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Towns 鄉鎮市區
      summary: 刪除鄉鎮市區
      operationId: delete_town_api_v1_towns__town_code__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: town_code
        in: path
        required: true
        schema:
          type: string
          title: Town Code
      - name: force
        in: query
        required: false
        schema:
          type: boolean
          description: 若仍有景點使用此行政區，是否強制刪除（景點的行政區欄位會被清空）
          default: false
          title: Force
        description: 若仍有景點使用此行政區，是否強制刪除（景點的行政區欄位會被清空）
      responses:
        '204':
          description: Successful Response
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CityCreate:
      properties:
        code:
          type: string
          maxLength: 10
          minLength: 1
          pattern: ^[0-9A-Za-z]+$
          title: Code
          examples:
          - '99001'
        name:
          type: string
          maxLength: 20
          minLength: 1
          title: Name
          examples:
          - 測試市
      type: object
      required:
      - code
      - name
      title: CityCreate
    CityOut:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        town_count:
          type: integer
          title: Town Count
          default: 0
        attraction_count:
          type: integer
          title: Attraction Count
          default: 0
      type: object
      required:
      - code
      - name
      title: CityOut
...
    CityUpdate:
      properties:
        name:
          type: string
          maxLength: 20
          minLength: 1
          title: Name
          examples:
          - 測試新市
      type: object
      required:
      - name
      title: CityUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Message:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: Message
...
    Page_CityOut_:
      properties:
        total:
          type: integer
          title: Total
          description: 符合條件的總筆數
        page:
          type: integer
          title: Page
          description: 目前頁碼（從 1 開始）
        page_size:
          type: integer
          title: Page Size
          description: 每頁筆數
        pages:
          type: integer
          title: Pages
          description: 總頁數
        items:
          items:
            $ref: '#/components/schemas/CityOut'
          type: array
          title: Items
      type: object
      required:
      - total
      - page
      - page_size
      - pages
      - items
      title: Page[CityOut]
...
    Page_TownOut_:
      properties:
        total:
          type: integer
          title: Total
          description: 符合條件的總筆數
        page:
          type: integer
          title: Page
          description: 目前頁碼（從 1 開始）
        page_size:
          type: integer
          title: Page Size
          description: 每頁筆數
        pages:
          type: integer
          title: Pages
          description: 總頁數
        items:
          items:
            $ref: '#/components/schemas/TownOut'
          type: array
          title: Items
      type: object
      required:
      - total
      - page
      - page_size
      - pages
      - items
      title: Page[TownOut]
...
    TownCreate:
      properties:
        code:
          type: string
          maxLength: 12
          minLength: 1
          pattern: ^[0-9A-Za-z]+$
          title: Code
          examples:
          - '99001010'
        name:
          type: string
          maxLength: 20
          minLength: 1
          title: Name
          examples:
          - 測試區
        city_code:
          type: string
          title: City Code
          examples:
          - '99001'
      type: object
      required:
      - code
      - name
      - city_code
      title: TownCreate
    TownOut:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        city_code:
          type: string
          title: City Code
        city_name:
          anyOf:
          - type: string
          - type: 'null'
          title: City Name
        attraction_count:
          type: integer
          title: Attraction Count
          default: 0
      type: object
      required:
      - code
      - name
      - city_code
      title: TownOut
...
    TownUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 20
            minLength: 1
          - type: 'null'
          title: Name
        city_code:
          anyOf:
          - type: string
          - type: 'null'
          title: City Code
      type: object
      title: TownUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: 寫入操作需帶入 API Key（預設為 `ntub-iot-2026`，可用環境變數 `API_KEY` 修改）
      in: header
      name: X-API-Key
...
servers:
- url: http://127.0.0.1:8000
  description: 本機開發伺服器